trainer/requirements.txt
accelerate==0.27.2
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
appdirs==1.4.4
archspec @ file:///croot/archspec_1697725767277/work
async-timeout==4.0.3
attrs==23.2.0
azure-core==1.30.1
azure-storage-blob==12.19.1
bitsandbytes==0.43.0
boltons @ file:///croot/boltons_1677628692245/work
Brotli @ file:///tmp/abs_ecyw11_7ze/croots/recipe/brotli-split_1659616059936/work
certifi @ file:///croot/certifi_1707229174982/work/certifi
cffi @ file:///croot/cffi_1700254295673/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.7
conda @ file:///croot/conda_1708369113911/work
conda-content-trust @ file:///croot/conda-content-trust_1693490622020/work
conda-libmamba-solver @ file:///croot/conda-libmamba-solver_1706733287605/work/src
conda-package-handling @ file:///croot/conda-package-handling_1690999929514/work
conda_package_streaming @ file:///croot/conda-package-streaming_1690987966409/work
cryptography @ file:///croot/cryptography_1707523700518/work
datasets==2.18.0
deepspeed==0.14.0
dill==0.3.8
distro @ file:///croot/distro_1701455004953/work
docker-pycreds==0.4.0
docstring-parser==0.15
einops==0.7.0
filelock==3.9.0
flash-attn==2.5.6
frozenlist==1.4.1
fsspec==2024.2.0
gitdb==4.0.11
GitPython==3.1.42
hjson==3.1.0
huggingface-hub==0.21.4
idna @ file:///croot/idna_1666125576474/work
isodate==0.6.1
Jinja2==3.1.2
joblib==1.3.2
jsonpatch @ file:///tmp/build/80754af9/jsonpatch_1615747632069/work
jsonpointer==2.1
libmambapy @ file:///croot/mamba-split_1704219408234/work/libmambapy
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
menuinst @ file:///croot/menuinst_1706732933928/work
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
networkx==3.2.1
ninja==1.11.1.1
numpy==1.26.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.1.105
nvidia-nvtx-cu12==12.1.105
packaging @ file:///croot/packaging_1693575174725/work
pandas==2.2.1
peft==0.9.0
pillow==10.2.0
platformdirs @ file:///croot/platformdirs_1692205439124/work
pluggy @ file:///tmp/build/80754af9/pluggy_1648024709248/work
protobuf==4.25.3
psutil==5.9.8
py-cpuinfo==9.0.0
pyarrow==15.0.1
pyarrow-hotfix==0.6
pycosat @ file:///croot/pycosat_1696536503704/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==2.6.3
pydantic_core==2.16.3
Pygments==2.17.2
pynvml==11.5.0
PySocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests @ file:///croot/requests_1707355572290/work
rich==13.7.1
ruamel.yaml @ file:///croot/ruamel.yaml_1666304550667/work
ruamel.yaml.clib @ file:///croot/ruamel.yaml.clib_1666302247304/work
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
sentencepiece==0.2.0
sentry-sdk==1.41.0
setproctitle==1.3.3
shtab==1.7.1
six==1.16.0
smmap==5.0.1
sympy==1.12
threadpoolctl==3.3.0
tokenizers==0.15.2
torch==2.2.1+cu121
torchaudio==2.2.1+cu121
torchvision==0.17.1+cu121
tqdm @ file:///croot/tqdm_1679561862951/work
transformers==4.38.2
triton==2.2.0
trl==0.7.11
truststore @ file:///croot/truststore_1695244293384/work
typing_extensions==4.8.0
tyro==0.7.3
tzdata==2024.1
urllib3 @ file:///croot/urllib3_1707770551213/work
wandb==0.16.4
xxhash==3.4.1
yarl==1.9.4
zstandard @ file:///croot/zstandard_1677013143055/work